--- core/server/routes/admin.js
@@ -1,14 +1,0 @@
-var admin       = require('../controllers/admin'),
-    express     = require('express'),
-
-    adminRoutes;
-
-adminRoutes = function () {
-    var router = express.Router();
-
-    router.get('*', admin.index);
-
-    return router;
-};
-
-module.exports = adminRoutes;
