+++ core/server/routes/index.js
@@ -1,4 +1,3 @@
 module.exports = {
-    admin: require('./admin'),
     frontend: require('./frontend')
 };
